requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gitdb==4.0.5
GitPython==3.1.9
google-auth==1.22.0
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.11.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==0.9.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.91
sentry-sdk==0.18.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
torch==1.6.0
tornado==6.0.4
tqdm==4.50.0
traitlets==5.0.4
transformers==3.3.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wandb==0.10.4
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yarl==1.6.0
zipp==3.3.0
